Rename data package to database (#9727)
